Folder change and added uart ax support
