+++ .cproject
@@ -33,8 +33,9 @@
 								<option id="de.innot.avreclipse.compiler.option.def.482357620" name="Define Syms (-D)" superClass="de.innot.avreclipse.compiler.option.def" useByScannerDiscovery="false" valueType="definedSymbols">
 									<listOptionValue builtIn="false" value="F_CPU=10000000"/>
 								</option>
-								<option id="de.innot.avreclipse.compiler.option.incpath.835990303" superClass="de.innot.avreclipse.compiler.option.incpath" useByScannerDiscovery="false" valueType="includePath">
-									<listOptionValue builtIn="false" value="&quot;${workspace_loc:/${ProjName}/inc}&quot;"/>
+								<option id="de.innot.avreclipse.compiler.option.incpath.835990303" name="Include Paths (-I)" superClass="de.innot.avreclipse.compiler.option.incpath" useByScannerDiscovery="false" valueType="includePath">
+									<listOptionValue builtIn="false" value="&quot;${workspace_loc:/${ProjName}/Source/internal/inc}&quot;"/>
+									<listOptionValue builtIn="false" value="&quot;${workspace_loc:/${ProjName}/Source/external/inc}&quot;"/>
 								</option>
 								<inputType id="de.innot.avreclipse.compiler.winavr.input.460344657" name="C Source Files" superClass="de.innot.avreclipse.compiler.winavr.input"/>
 							</tool>
