+++ Debug/sources.mk
@@ -20,6 +20,7 @@ C_DEPS :=
 
 # Every subdirectory with source files must be described here
 SUBDIRS := \
+Source/external/src \
+Source/internal/src \
 . \
-src \
