+++ Debug/src/actuator.d
@@ -1,13 +1,13 @@
 src/actuator.o: ../src/actuator.c \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/actuator.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/can.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h
+ /home/silardg/Development/Robotics/Software/Glavna/inc/actuator.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/can.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/actuator.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/actuator.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/can.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/can.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h:
