+++ Debug/src/can.d
@@ -1,4 +1,4 @@
 src/can.o: ../src/can.c \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/can.h
+ /home/silardg/Development/Robotics/Software/Glavna/inc/can.h
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/can.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/can.h:
