+++ Debug/src/gpio.d
@@ -1,7 +1,7 @@
 src/gpio.o: ../src/gpio.c \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/list_generic.h
+ /home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/list_generic.h
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/list_generic.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/list_generic.h:
