+++ Debug/src/odometry.d
@@ -1,10 +1,10 @@
 src/odometry.o: ../src/odometry.c \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/odometry.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/system.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/can.h
+ /home/silardg/Development/Robotics/Software/Glavna/inc/odometry.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/system.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/can.h
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/odometry.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/odometry.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/system.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/system.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/can.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/can.h:
