+++ Debug/src/sensors.d
@@ -1,10 +1,10 @@
 src/sensors.o: ../src/sensors.c \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/sensors.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/odometry.h
+ /home/silardg/Development/Robotics/Software/Glavna/inc/sensors.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/odometry.h
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/sensors.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/sensors.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/odometry.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/odometry.h:
