+++ Debug/src/system.d
@@ -1,25 +1,25 @@
 src/system.o: ../src/system.c \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/system.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/list_generic.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/can.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/usart.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/actuator.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/sensors.h
+ /home/silardg/Development/Robotics/Software/Glavna/inc/system.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/list_generic.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/can.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/usart.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/actuator.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h \
+ /home/silardg/Development/Robotics/Software/Glavna/inc/sensors.h
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/system.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/system.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/list_generic.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/list_generic.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/can.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/can.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/usart.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/usart.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/actuator.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/actuator.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/gpio.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/gpio.h:
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/sensors.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/sensors.h:
