+++ Debug/src/usart.d
@@ -1,4 +1,4 @@
 src/usart.o: ../src/usart.c \
- /home/silardg/Development/Robotics/Software/MainBoard/inc/usart.h
+ /home/silardg/Development/Robotics/Software/Glavna/inc/usart.h
 
-/home/silardg/Development/Robotics/Software/MainBoard/inc/usart.h:
+/home/silardg/Development/Robotics/Software/Glavna/inc/usart.h:
